packet: abstract DNS
Route53 is needed to deploy a cluster in Packet, this is a limitation as some
users don't have an AWS account or use another DNS provider.

This commit removes the specific Route53 logic and abstracts the DNS provider by
using a module that receives a list of DNS entries to be created.

variable "entries" {
  type = list(
    object({
      name    = string
      type    = string
      ttl     = number
      records = list(string)
    })
  )
}

The user should change the module to be used by editing the source field of the
DNS module in the packet/flatcar-linux/kubernetes/dns.tf file.

Unfortunately Terraform doesn't support variables in the source of a module
(https://github.com/hashicorp/terraform/issues/1439) so it cannot be configurable.

The default source is to the manual provider for now.