packet: add DNS manual provider
It is possible that in some situations user's DNS provider is not supoported or
that a manual procedure needs to be done to create the entries, for instance, to
request it to a different team.  In order to support those scenarios this commit
implements a manual DNS provider that saves the list of entries to be created in
a file, the user can create those manually and then continue with the deployment.

In order to know the values for the DNS entries the controller nodes need to be
created, in order to do so the -target options of Terraform apply can be used:

`terraform apply -target=module.<controller-module-name>.module.dns-entries`

This command will create a `dns-entries.txt` file, once the user has created
those entries manually, it can continue the deployment by executing:

`terraform apply`